chore: remove code from CLI package
