--- packages/cli/src/lib/parseOptions.ts
@@ -1,32 +1,0 @@
-/*
- * Copyright 2023 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export function parseOptions(optionStrings: string[]): Record<string, string> {
-  const options: Record<string, string> = {};
-
-  for (const str of optionStrings) {
-    const [key] = str.split('=', 1);
-    const value = str.slice(key.length + 1);
-    if (!key || str[key.length] !== '=') {
-      throw new Error(
-        `Invalid option '${str}', must be of the format <key>=<value>`,
-      );
-    }
-    options[key] = value;
-  }
-
-  return options;
-}
