--- packages/cli/templates/pulumi/.dockerignore
@@ -1,6 +1,0 @@
-dist-types
-node_modules
-packages/*/dist
-packages/*/node_modules
-plugins/*/dist
-plugins/*/node_modules
