--- packages/cli/templates/pulumi/Dockerfile
@@ -1,92 +1,0 @@
-# Stage 1 - Create yarn install skeleton layer
-FROM --platform=linux/amd64 node:16-bullseye-slim AS packages
-
-WORKDIR /app
-COPY yarn.lock ./
-COPY package.json ./package.json
-
-COPY packages packages
-COPY examples examples
-
-# Comment this out if you don't have any internal plugins
-# COPY plugins plugins
-
-RUN find packages \! -name "package.json" -mindepth 2 -maxdepth 2 -exec rm -rf {} \+
-
-# Stage 2 - Install dependencies and build packages
-FROM --platform=linux/amd64 node:16-bullseye-slim AS build
-
-# install sqlite3 dependencies
-RUN --mount=type=cache,target=/var/cache/apt,sharing=locked \
-    --mount=type=cache,target=/var/lib/apt,sharing=locked \
-    apt-get update && \
-    apt-get install -y --no-install-recommends libsqlite3-dev python3 build-essential && \
-    yarn config set python /usr/bin/python3
-
-USER node
-WORKDIR /app
-
-COPY --from=packages --chown=node:node /app .
-
-# Stop cypress from downloading it's massive binary.
-ENV CYPRESS_INSTALL_BINARY=0
-RUN --mount=type=cache,target=/home/node/.cache/yarn,sharing=locked,uid=1000,gid=1000 \
-    yarn install --network-timeout 600000
-
-COPY --chown=node:node . .
-
-RUN yarn tsc
-RUN yarn --cwd packages/backend build
-
-# If you have not yet migrated to package roles, use the following command instead:
-# RUN yarn --cwd packages/backend backstage-cli backend:bundle --build-dependencies
-
-RUN mkdir packages/backend/dist/skeleton packages/backend/dist/bundle \
-    && tar xzf packages/backend/dist/skeleton.tar.gz -C packages/backend/dist/skeleton \
-    && tar xzf packages/backend/dist/bundle.tar.gz -C packages/backend/dist/bundle
-
-# Stage 3 - Build the actual backend image and install production dependencies
-FROM --platform=linux/amd64 node:16-bullseye-slim
-
-# Install sqlite3 dependencies. You can skip this if you don't use sqlite3 in the image,
-# in which case you should also move better-sqlite3 to "devDependencies" in package.json.
-RUN --mount=type=cache,target=/var/cache/apt,sharing=locked \
-    --mount=type=cache,target=/var/lib/apt,sharing=locked \
-    apt-get update && \
-    apt-get install -y --no-install-recommends libsqlite3-dev python3 build-essential && \
-    yarn config set python /usr/bin/python3
-
-# From here on we use the least-privileged `node` user to run the backend.
-USER node
-
-# This should create the app dir as `node`.
-# If it is instead created as `root` then the `tar` command below will fail: `can't create directory 'packages/': Permission denied`.
-# If this occurs, then ensure BuildKit is enabled (`DOCKER_BUILDKIT=1`) so the app dir is correctly created as `node`.
-WORKDIR /app
-
-# Copy the install dependencies from the build stage and context
-COPY --from=build --chown=node:node /app/yarn.lock /app/packages/backend/dist/skeleton/ ./
-COPY package.json ./package.json
-COPY examples/ ./examples
-
-RUN cat package.json
-
-RUN --mount=type=cache,target=/home/node/.cache/yarn,sharing=locked,uid=1000,gid=1000 \
-    yarn install --network-timeout 600000
-
-# Copy the built packages from the build stage
-COPY --from=build --chown=node:node /app/packages/backend/dist/bundle/ ./
-
-# Copy any other files that we need at runtime
-COPY --chown=node:node app-config.yaml ./
-
-# To make sure we are not running Docker in Docker for the deployment
-RUN echo "techdocs:" >> app-config.deployment.yaml
-RUN echo "    generator:" >> app-config.deployment.yaml
-RUN echo "        runIn: 'local'" >> app-config.deployment.yaml
-
-
-# This switches many Node.js dependencies to development mode.
-ENV NODE_ENV development
-
-CMD ["node", "packages/backend", "--config",  "app-config.deployment.yaml", "--config", "app-config.yaml"]
